@@ -2,10 +2,9 @@ name: soundcode
 description: SoundCode flutter plugin project.
 version: 0.0.1
 author:
-homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <4.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
