@@ -2,23 +2,23 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 
 class SoundCode {
-
-  Function(List<int>) _callback;
-  Function() _error;
+  Function(List<int>)? _callback;
+  Function()? _error;
   static const MethodChannel _channel = const MethodChannel('com.cifrasoft.soundcode');
 
   static final SoundCode _singleton = SoundCode._internal();
   factory SoundCode() => _singleton;
 
-  SoundCode._internal(){
+  SoundCode._internal() {
     print("SoundCode Plugin");
     _channel.setMethodCallHandler((call) => receiveData(call));
   }
 
-  setCallback({Function(List<int>) callback}){
+  setCallback({Function(List<int>)? callback}) {
     this._callback = callback;
   }
-  setErrorCallback({Function() error}){
+
+  setErrorCallback({Function()? error}) {
     this._error = error;
   }
 
@@ -28,33 +28,28 @@ class SoundCode {
     return result;
   }
 
-  Future<String>  start() async {
+  Future<String> start() async {
     final String result = await _channel.invokeMethod('start');
     return result;
   }
 
-  Future<String>  stop() async {
+  Future<String> stop() async {
     final String result = await _channel.invokeMethod('stop');
     return result;
   }
 
   //from java
   Future<void> receiveData(MethodCall call) async {
-    switch(call.method) {
+    switch (call.method) {
       case "onDetectedId":
         print("onDetectedId ");
-        if(this._callback!=null) {
-          var data = List<int>.from(call.arguments);
-          _callback(data);
-        }
+        var data = List<int>.from(call.arguments as List<dynamic>);
+        _callback?.call(data);
         break;
       case "onAudioInitFailed":
         print("onAudioInitFailed ");
-        if(this._error!=null) {
-          _error();
-        }
+        _error?.call();
         break;
     }
   }
-
 }
